Rename mix_3d files, add geo_act params, revise scaling_mode selection
