+++ scripts/mixed_3d/mpcno_geo_mixed_3d_parallel_train.py
@@ -11,12 +11,12 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 import numpy as np
 from timeit import default_timer
 
-from pcno_geo_mixed_3d_helper import gen_data_tensors
+from scripts.mixed_3d.mpcno_geo_mixed_3d_helper import gen_data_tensors
 
 sys.path.append(str(Path(__file__).parent.parent))
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
 
-from pcno.pcno_geo import compute_Fourier_modes, PCNO, PCNO_train_parallel
+from pcno.mpcno import compute_Fourier_modes, MPCNO, MPCNO_train_parallel
 
 torch.set_printoptions(precision=16)
 
@@ -63,6 +63,7 @@ def train_ddp(rank, local_rank, world_size, args):
     ndim = 3
     layers = [int(size) for size in args.layer_sizes.split(",")]
     act = args.act
+    geo_act = args.geo_act
     to_divide_factor = args.to_divide_factor
     mesh_type = args.mesh_type
     n_train = args.n_train
@@ -137,13 +138,15 @@ def train_ddp(rank, local_rank, world_size, args):
 
     modes = compute_Fourier_modes(ndim, [k_max, k_max, k_max], Ls)
     modes = torch.tensor(modes, dtype=torch.float, device=f'cuda:{local_rank}')
-    model = PCNO(ndim, modes, nmeasures=1,
+    model = MPCNO(ndim, modes, nmeasures=1,
                 layer_selection = layer_selection,
                 layers=layers,
                 fc_dim=128,
                 in_dim=x_train.shape[-1], out_dim=y_train.shape[-1],
                 inv_L_scale_hyper = [train_inv_L_scale, 0.5, 2.0],
                 act = act,
+                geo_act = geo_act,
+                scaling_mode='sqrt_inv',
                 ).to(local_rank)
 
     # Wrap the model with DDP
@@ -173,7 +176,7 @@ def train_ddp(rank, local_rank, world_size, args):
                         }
 
 
-    train_rel_l2_losses, test_rel_l2_losses, test_l2_losses = PCNO_train_parallel(
+    train_rel_l2_losses, test_rel_l2_losses, test_l2_losses = MPCNO_train_parallel(
         x_train, aux_train, y_train, x_test, aux_test, y_test, config, ddp_model, rank=rank, local_rank = local_rank, world_size=world_size, save_model_name="./PCNO_parallel_mixed_3d_model"
     )
     
@@ -198,6 +201,7 @@ def main():
     parser.add_argument('--n_train', type=int, default=900)
     parser.add_argument('--n_test', type=int, default=100)
     parser.add_argument('--act', type=str, default="gelu")
+    parser.add_argument('--geo_act', type=str, default="softsign")
     parser.add_argument('--scale', type=float, default=0.0)
     parser.add_argument("--layer_sizes", type=str, default="64,64,64,64,64,64")
