+++ scripts/mixed_3d/mpcno_parallel_train.sh
