+++ scripts/mixed_3d/mpcno_plot_results.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
-#SBATCH -o logs/PCNO_plot_results_vertex_centered.out
+#SBATCH -o logs/MPCNO_plot_results_vertex_centered.out
 #SBATCH --qos=low
 #SBATCH -p C064M1024G
-#SBATCH -J PCNO_plot_results_vertex_centered
+#SBATCH -J MPCNO_plot_results_vertex_centered
 #SBATCH --nodes=1 
 #SBATCH --ntasks=1
 #SBATCH --cpus-per-task=6
@@ -10,4 +10,4 @@
 
 module load conda
 source activate pytorch
-python pcno_plot_results.py 
+python mpcno_plot_results.py
