+++ scripts/mixed_3d/mpcno_preprocess_data.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
-#SBATCH -o logs/PCNO_preprocess_data_vertex_centered.out
+#SBATCH -o logs/MPCNO_preprocess_data_vertex_centered.out
 #SBATCH --qos=low
 #SBATCH -p C064M1024G
-#SBATCH -J PCNO_preprocess_data_vertex_centered
+#SBATCH -J MPCNO_preprocess_data_vertex_centered
 #SBATCH --nodes=1 
 #SBATCH --ntasks=1
 #SBATCH --cpus-per-task=32
@@ -10,5 +10,5 @@
 
 module load conda
 source activate pytorch
-python pcno_geo_mixed_3d_preprocess_data.py --n_each 2000 \
+python mpcno_geo_mixed_3d_preprocess_data.py --n_each 2000 \
                                  --mesh_type "vertex_centered"  # "cell_centered" , "vertex_centered"
